// src/app/trainee-biodata/sailor-biodata-tab-layout/models/ParentRelative.ts
export interface ParentRelative {
    parentRelativeId: number;
    traineeId: number;
    relationTypeId: number;
    relationType: string;
    maritalStatus: string;
    deadStatusValue: string;
    nationality: string;
    religion: string;
    caste: string;
    occupation: string;
    previousOccupation: string;
    division: string;
    district: string;
    thana: string;
    defenseType: string;
    rank: string;
    secondNationality: string;
    nameInFull: string;
    dateOfBirth: Date;
    maritalStatusId: number;
    dteOfMarriage: Date;
    age: number;
    deadStatus: number;
    dateOfExpiry: Date;
    nationalityId: number;
    religionId: number;
    casteId: number;
    occupationId: number;
    occupationDetail: string;
    educationQualification: string;
    previousOccupationId: number;
    monthlyIncome: string;
    divisionId: number;
    districtId: number;
    thanaId: number;
    postCode: string;
    presentAddress: string;
    parmanentAddress: string;
    mobile: string;
    email: string;
    remarks: string;
    isDefenceJobExperience: string;
    defenseTypeId: number;
    rankId: number;
    retiredDate: Date;
    placeOfBirth: string;
    dualNationality: string;
    secondNationalityId: number;
    reasonOfMigration: string;
    migrationDate: Date;
    additionalInformation: string;
    status: number;
    menuPosition: number;
    isActive: boolean;  
} 